devop: move buy providers back to components
