+++ src/modules/buy-sell/components/BuyProviderComponent.vue
@@ -2,424 +2,234 @@
   <div class="py-8 px-8 pt-3">
     <div
       class="d-flex align-center textDark--text mb-10 cursor--pointer"
-      @click="$emit('close')"
+      @click="step = 0"
     >
       <v-icon color="textDark">mdi-arrow-left mr-4</v-icon>
-      <div class="mew-heading-2">Select provider</div>
+      <div class="mew-heading-2">Select a provider to buy</div>
     </div>
-
-    <!-- ============================================================== -->
-    <!-- Moonpay -->
-    <!-- ============================================================== -->
-    <div v-if="!hideMoonpay" class="section-block pa-5 mb-6">
-      <img
-        class="provider-logo"
-        src="@/modules/buy-sell/assets/moonpay-logo.svg"
-        height="18"
-      />
-      <div class="mb-3">
-        <div class="d-flex mb-1 align-center">
-          <div class="d-flex mew-heading-3 textDark--text mr-1">
-            {{ moonpayQuote.cryptoToFiat }}
-            <span class="mew-heading-3 pl-1">{{ selectedCryptoName }}</span>
+    <div class="mew-heading-2 mb-10">
+      Spending {{ formattedFiat }} to buy {{ selectedCurrency.symbol }} on
+      {{ network.type.name_long }} Network
+    </div>
+    <div>
+      <div
+        v-for="(provider, idx) in buyQuote"
+        :key="provider.provider + idx"
+        class="section-block ripple pa-5 mb-6"
+        @click="buyProvider(provider)"
+      >
+        <div v-if="idx === 0" class="best-rate">Best Rate</div>
+        <div>
+          <div class="d-flex mew-heading-3">
+            {{ provider.crypto_amount }}
+            <div class="d-flex align-center">
+              <span class="mew-heading-3 pl-1 mr-1">{{
+                provider.crypto_currency
+              }}</span>
+              <v-tooltip location="bottom" min-width="200px">
+                <template #activator="{ props }">
+                  <v-icon
+                    v-bind="props"
+                    color="grey-lighten-1"
+                    size="x-small"
+                    class="cursor-pointer"
+                  >
+                    mdi-information
+                  </v-icon>
+                </template>
+                <div class="elevated-box pa-3">
+                  {{ generateFeeLabel(provider) }}
+                  <br />
+                  <br />
+                  <br />
+                  {{ generateLimits(provider.provider) }}
+                  <br />
+                  {{ generateLimits(provider.provider, false) }}
+                </div>
+              </v-tooltip>
+            </div>
           </div>
-          <mew-tooltip>
-            <template #contentSlot>
-              <div>
-                {{ moonpayQuote.includesFeeText }}
-                <br />
-                <br />
-                {{ moonpayQuote.networkFeeText }}
-                <br />
-                <br />
-                {{ moonpayQuote.dailyLimit }}
-                <br />
-                {{ moonpayQuote.monthlyLimit }}
+          <div class="d-flex align-center justify-space-between mt-3">
+            <img
+              :src="parseProviderLogo(provider)"
+              :alt="provider.provider + ' logo'"
+              width="100"
+            />
+            <div class="d-flex flex-column align-center mb-1">
+              <div
+                :class="[
+                  'd-flex align-center mb-1 justify-end',
+                  $vuetify.breakpoint.smAndDown ? 'flex-wrap' : ''
+                ]"
+                :style="
+                  $vuetify.breakpoint.smAndDown
+                    ? 'width : 150px;'
+                    : 'width: 100%'
+                "
+              >
+                <img
+                  v-for="(logo, providerIdx) in parsePaymentMethods(
+                    provider.payment_methods
+                  )"
+                  :key="providerIdx + logo"
+                  :src="logo"
+                  :alt="provider.provider + ' payment method'"
+                  width="40"
+                  class="mr-1"
+                />
               </div>
-            </template>
-          </mew-tooltip>
-        </div>
-      </div>
-
-      <div class="d-flex align-center mb-1">
-        <img
-          src="@/assets/images/icons/moonpay/icon-visa.svg"
-          alt="Visa"
-          height="24"
-          class="mr-2"
-        />
-        <img
-          src="@/assets/images/icons/moonpay/icon-master.svg"
-          alt="Master"
-          height="24"
-          class="mr-2"
-        />
-        <img
-          src="@/assets/images/icons/moonpay/icon-apple-pay.svg"
-          alt="applepay"
-          height="24"
-          class="mr-2"
-        />
-        <img
-          src="@/assets/images/icons/moonpay/icon-paypal-logo.svg"
-          alt="paypal"
-          height="24"
-          class="mr-2"
-        />
-        <img
-          v-if="isEUR"
-          src="@/assets/images/icons/moonpay/icon-bank.svg"
-          alt="Bank"
-          height="24"
-        />
-      </div>
-      <div class="mew-label mb-5">
-        {{ paymentOptionString }}
-      </div>
-      <div>
-        <mew-button
-          btn-size="large"
-          btn-style="light"
-          color-theme="buttonGrayLight"
-          has-full-width
-          :is-valid-address-func="isValidToAddress"
-          :title="moonpayBtnTitle"
-          @click.native="buy"
-        />
-      </div>
-    </div>
-
-    <!-- ============================================================== -->
-    <!-- Simplex -->
-    <!-- ============================================================== -->
-    <div v-if="!hideSimplex" class="section-block pa-5">
-      <div v-if="!loading" class="mb-3">
-        <div class="d-flex mb-1 align-center">
-          <div class="d-flex mew-heading-3 textDark--text mr-1">
-            {{ simplexQuote.crypto_amount }}
-            <span class="mew-heading-3 pl-1">{{ selectedCryptoName }}</span>
+              <div class="mew-label d-none d-sm-block">
+                {{ parsePaymentMethods(provider.payment_methods, true) }}
+              </div>
+            </div>
           </div>
-          <mew-tooltip>
-            <template #contentSlot>
-              <div>
-                {{ moonpayQuote.includesFeeText }}
-                <br />
-                <br />
-                {{ moonpayQuote.networkFeeText }}
-                <br />
-                <br />
-                {{ moonpayQuote.dailyLimit }}
-                <br />
-                {{ moonpayQuote.monthlyLimit }}
-              </div>
-            </template>
-          </mew-tooltip>
         </div>
-      </div>
-
-      <div v-else class="mb-3">
-        <v-skeleton-loader type="heading" class="mb-1" />
-        <v-skeleton-loader max-width="200px" type="heading" />
-      </div>
-
-      <div class="d-flex align-center justify-space-between">
-        <div class="d-flex align-start mb-1">
-          <img
-            src="@/assets/images/icons/moonpay/icon-visa.svg"
-            alt="Visa"
-            height="24"
-            class="mr-2"
-          />
-          <img
-            src="@/assets/images/icons/moonpay/icon-master.svg"
-            alt="Master"
-            height="24"
-            class="mr-2"
-          />
-        </div>
-        <img
-          class="provider-logo"
-          src="@/assets/images/icons/icon-simplex.svg"
-          alt="simplex"
-          height="28"
-        />
-      </div>
-      <div class="mew-label mb-5">Visa, Mastercard</div>
-      <div>
-        <mew-button
-          btn-size="large"
-          btn-style="light"
-          color-theme="buttonGrayLight"
-          has-full-width
-          :title="simplexBtnTitle"
-          @click.native="openSimplex"
-        />
       </div>
     </div>
-    <!-- ============================================================== -->
-    <!-- Topper -->
-    <!-- ============================================================== -->
-    <div v-if="!hideTopper" class="section-block pa-5">
-      <div v-if="!loading" class="mb-3">
-        <div class="d-flex mb-1 align-center">
-          <div class="d-flex mew-heading-3 textDark--text mr-1">
-            {{ topperQuote.cryptoToFiat }}
-            <span class="mew-heading-3 pl-1">{{ selectedCryptoName }}</span>
-          </div>
-          <mew-tooltip>
-            <template #contentSlot>
-              <div>
-                {{ topperQuote.includesFeeText }}
-                <br />
-                <br />
-                {{ topperQuote.networkFeeText }}
-                <br />
-                <br />
-                {{ topperQuote.dailyLimit }}
-                <br />
-                {{ topperQuote.monthlyLimit }}
-              </div>
-            </template>
-          </mew-tooltip>
-        </div>
-      </div>
-
-      <div v-else class="mb-3">
-        <v-skeleton-loader type="heading" class="mb-1" />
-        <v-skeleton-loader max-width="200px" type="heading" />
-      </div>
-
-      <div class="d-flex align-center justify-space-between">
-        <div class="d-flex align-start mb-1">
-          <img
-            src="@/assets/images/icons/moonpay/icon-visa.svg"
-            alt="Visa"
-            height="24"
-            class="mr-2"
-          />
-          <img
-            src="@/assets/images/icons/moonpay/icon-master.svg"
-            alt="Master"
-            height="24"
-            class="mr-2"
-          />
-          <img
-            src="@/assets/images/icons/moonpay/icon-apple-pay.svg"
-            alt="ApplePay"
-            height="24"
-            class="mr-2"
-          />
-          <img
-            src="@/assets/images/icons/moonpay/icon-google-pay-logo.svg"
-            alt="googlePay"
-            height="35"
-            class="mr-2"
-            style="margin-top: -5px"
-          />
-          <img
-            src="@/assets/images/icons/moonpay/icon-pix-logo.svg"
-            alt="pixpay"
-            height="24"
-            class="mr-2"
-          />
-        </div>
-        <img
-          class="provider-logo"
-          src="@/assets/images/icons/icon-topper.png"
-          alt="topper"
-          height="28"
-        />
-      </div>
-      <div class="mew-label mb-5">
-        Visa, Mastercard, Apple Pay, Google Pay, Pix
-      </div>
-      <div>
-        <mew-button
-          btn-size="large"
-          btn-style="light"
-          color-theme="buttonGrayLight"
-          has-full-width
-          :title="topperBtnTitle"
-          @click.native="openTopper"
-        />
-      </div>
-    </div>
-
-    <div v-if="hideSimplex && hideMoonpay" class="section-block pa-5">
-      <div class="d-flex mew-heading-3 textDark--text">No Providers found</div>
+    <div class="pt-2 text-center mew-label">
+      Fees, availability, and purchase limits vary between providers, you can
+      check their quotes and select one that works for you.
     </div>
   </div>
 </template>
 
 <script>
-import MultiCoinValidator from 'multicoin-address-validator';
-import { mapGetters, mapState } from 'vuex';
-import { isEmpty } from 'lodash';
-
-import { ERROR, Toast } from '@/modules/toast/handler/handlerToast';
-import handlerAnalytics from '@/modules/analytics-opt-in/handlers/handlerAnalytics.mixin';
-import { BUY_SELL } from '@/modules/analytics-opt-in/handlers/configs/events';
-import BigNumber from 'bignumber.js';
+import { mapGetters } from 'vuex';
+import { currencySymbols } from './tokenList';
+import visaLogo from '@/assets/images/icons/moonpay/icon-visa.svg';
+import mastercardLogo from '@/assets/images/icons/moonpay/icon-master.svg';
+import bankLogo from '@/assets/images/icons/moonpay/icon-bank.svg';
+import applePayLogo from '@/assets/images/icons/moonpay/icon-apple-pay.svg';
+import paypalLogo from '@/assets/images/icons/moonpay/icon-paypal-logo.svg';
+import googlePayLogo from '@/assets/images/icons/moonpay/icon-google-pay-logo.svg';
+import pixLogo from '@/assets/images/icons/moonpay/icon-pix-logo.svg';
 
 export default {
   name: 'ModuleBuyEthProvider',
-  mixins: [handlerAnalytics],
   props: {
-    orderHandler: {
-      type: Object,
-      default: () => {}
-    },
-    selectedFiat: {
-      type: Object,
-      default: () => ({})
+    buyQuote: {
+      type: Array,
+      default: () => []
     },
     selectedCurrency: {
       type: Object,
       default: () => ({})
     },
-    moonpayQuote: {
-      type: Object,
-      default: () => ({})
+    buyProvider: {
+      type: Function,
+      default: () => {}
     },
-    simplexQuote: {
-      type: Object,
-      default: () => ({})
-    },
-    topperQuote: {
+    selectedFiat: {
       type: Object,
       default: () => ({})
     }
   },
   data() {
-    return {
-      loading: false
-    };
+    return {};
   },
   computed: {
     ...mapGetters('global', ['network']),
-    ...mapState('wallet', ['address']),
-    selectedFiatName() {
-      return this.selectedFiat.name;
-    },
-    selectedCryptoName() {
-      return this.selectedCurrency.symbol;
+    formattedFiat() {
+      const symbol = currencySymbols[this.selectedFiat.name]
+        ? currencySymbols[this.selectedFiat.name]
+        : '';
+      return `${symbol}${this.buyQuote[0]?.fiat_amount || 0}`;
     },
     isEUR() {
-      return this.selectedFiatName === 'EUR' || this.selectedFiatName === 'GBP';
-    },
-    hideMoonpay() {
-      return isEmpty(this.moonpayQuote);
-    },
-    hideSimplex() {
-      return isEmpty(this.simplexQuote);
-    },
-    hideTopper() {
-      return BigNumber(this.topperQuote.cryptoToFiat).isZero();
-    },
-    simplexBtnTitle() {
-      return 'BUY WITH SIMPLEX';
-    },
-    moonpayBtnTitle() {
-      return 'BUY WITH MOONPAY';
-    },
-    topperBtnTitle() {
-      return 'BUY WITH TOPPER';
-    },
-    paymentOptionString() {
-      return `Visa, Mastercard, Apple Pay, Paypal${
-        this.isEUR ? ', Bank account' : ''
-      }`;
+      return this.selectedFiat.name === 'EUR';
     }
   },
   methods: {
-    isValidToAddress(address) {
-      return MultiCoinValidator.validate(address, this.selectedCurrency.symbol);
-    },
-    openTopper() {
-      if (!this.validToAddress()) return;
-      this.trackBuySell(BUY_SELL.BUY_W_TOPPER);
-      const param = {
-        fiatCurrency: this.selectedFiatName,
-        cryptoCurrency: this.selectedCryptoName,
-        requestedAmount: this.topperQuote.fiatAmount,
-        address: this.address
-      };
-      this.orderHandler
-        .getTopperUrl(param)
-        .then(res => {
-          window.open(res.url, '_blank');
-          this.trackBuySell(BUY_SELL.BUY_W_TOPPER_SUCCESS);
-          this.reset();
-          this.$emit('close');
-          this.$emit('reset');
-        })
-        .catch(err => {
-          this.trackBuySell(BUY_SELL.BUY_W_TOPPER_FAILED);
-          this.reset();
-          Toast(err, {}, ERROR);
-          this.$emit('close');
-          this.$emit('reset');
-        });
-    },
-    openSimplex() {
-      if (!this.validToAddress()) return;
-      this.trackBuySell(BUY_SELL.BUY_W_SIMPLEX);
-      this.orderHandler
-        .simplexBuy(
-          this.selectedCryptoName,
-          this.selectedFiatName,
-          this.moonpayQuote.fiatAmount,
-          this.address
-        )
-        .then(() => {
-          this.trackBuySell(BUY_SELL.BUY_W_SIMPLEX_SUCCESS);
-          this.reset();
-          this.$emit('close');
-          this.$emit('reset');
-        })
-        .catch(err => {
-          this.trackBuySell(BUY_SELL.BUY_W_SIMPLEX_FAILED);
-          this.reset();
-          Toast(err, {}, ERROR);
-          this.$emit('close');
-          this.$emit('reset');
-        });
-    },
-    reset() {
-      this.loading = true;
-      this.fetchData = {};
-    },
-    buy() {
-      if (!this.validToAddress()) return;
-      this.trackBuySell(BUY_SELL.BUY_W_MOONPAY);
-      this.orderHandler
-        .buy(
-          this.selectedCryptoName,
-          this.selectedFiatName,
-          this.moonpayQuote.fiatAmount,
-          this.address
-        )
-        .then(() => {
-          this.trackBuySell(BUY_SELL.BUY_W_MOONPAY_SUCCESS);
-          this.reset();
-          this.$emit('close');
-          this.$emit('reset');
-        })
-        .catch(err => {
-          this.trackBuySell(BUY_SELL.BUY_W_MOONPAY_FAILED);
-          this.reset();
-          Toast(err, {}, ERROR);
-          this.$emit('close');
-          this.$emit('reset');
-        });
-    },
-    validToAddress() {
-      if (isEmpty(this.address)) {
-        Toast('To address cannot be empty!', {}, ERROR);
-        this.$emit('close');
-        return false;
+    parsePaymentMethods(paymentMethods, label = false) {
+      const logos = [];
+      let paymentMethodsLabel = '';
+      if (
+        paymentMethods.includes('CREDIT_CARD') ||
+        paymentMethods.includes('DEBIT_CARD') ||
+        paymentMethods.includes('CARD')
+      ) {
+        logos.push(visaLogo);
+        logos.push(mastercardLogo);
+        paymentMethodsLabel += 'Visa, Mastercard';
       }
-      return true;
+
+      if (
+        (paymentMethods.includes('ACH') ||
+          paymentMethods.includes('ACH_BANK_ACCOUNT') ||
+          paymentMethods.includes('SEPA_OPEN_BANKING')) &&
+        this.isEUR
+      ) {
+        logos.push(bankLogo);
+        paymentMethodsLabel += `${
+          paymentMethodsLabel.length > 0 ? ', ' : ''
+        }Bank account`;
+      }
+
+      if (paymentMethods.includes('PAYPAL')) {
+        logos.push(paypalLogo);
+        paymentMethodsLabel += `${
+          paymentMethodsLabel.length > 0 ? ', ' : ''
+        }Paypal`;
+      }
+
+      if (paymentMethods.includes('APPLE_PAY')) {
+        logos.push(applePayLogo);
+        paymentMethodsLabel += `${
+          paymentMethodsLabel.length > 0 ? ', ' : ''
+        }Apple Pay`;
+      }
+
+      if (paymentMethods.includes('GOOGLE_PAY')) {
+        logos.push(googlePayLogo);
+        paymentMethodsLabel += `${
+          paymentMethodsLabel.length > 0 ? ', ' : ''
+        }Google Pay`;
+      }
+
+      if (paymentMethods.includes('PIX')) {
+        logos.push(pixLogo);
+        paymentMethodsLabel += `${
+          paymentMethodsLabel.length > 0 ? ', ' : ''
+        }Pix`;
+      }
+
+      return label ? paymentMethodsLabel : logos;
+    },
+    parseProviderLogo(provider) {
+      const providerLogos = {
+        SIMPLEX: require('@/assets/images/providers/icon-simplex.svg'),
+        MOONPAY: require('@/assets/images/providers/icon-moonpay.svg'),
+        TOPPER: require('@/assets/images/providers/icon-topper.svg'),
+        COINBASE: require('@/assets/images/providers/icon-coinbase-light.webp')
+      };
+      return providerLogos[provider.provider];
+    },
+    generateFeeLabel(provider) {
+      const feeLabel = {
+        MOONPAY: `Includes ${this.isEUR ? 0.7 : 4.99}% fee`,
+        SIMPLEX: 'Includes fee 5.25% fee',
+        TOPPER: 'Includes 4.65% fee. First transaction is free.',
+        COINBASE: 'Includes 2.5% fee.'
+      };
+
+      return feeLabel[provider.provider];
+    },
+    generateLimits(provider, daily = true) {
+      const dailyLimits = {
+        MOONPAY: 'Daily limit: $10,000',
+        SIMPLEX: 'Daily limit: $20,000',
+        TOPPER: 'Daily limit: $20,000',
+        COINBASE: 'Daily limit: $20,000'
+      };
+
+      const monthlyLimits = {
+        MOONPAY: 'Monthly limit: $50,000',
+        SIMPLEX: 'Monthly limit: $50,000',
+        TOPPER: 'Monthly limit: $50,000',
+        COINBASE: 'Monthly limit: $50,000'
+      };
+
+      return daily ? dailyLimits[provider] : monthlyLimits[provider];
     }
   }
 };
